# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Rest",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Autenticate"
                ],
                "summary": "Login user",
                "description": "Login user",
                "operationId": "2d0f20787e05d7face1de433b6727c52",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Fetch Users list",
                "description": "Fetch Users list",
                "operationId": "79e1348c56d8a16b5321753ffb26a294",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object",
                                    "example": [
                                        {
                                            "id": "9d12795b-138b-47c9-912e-01eff34cdaa5",
                                            "name": "User",
                                            "email": "user@example.com",
                                            "active": true,
                                            "created_at": "2024-09-22T17:38:29.000000Z",
                                            "updated_at": "2024-09-22T17:38:29.000000Z"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/users/{uuid}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Fetch User",
                "description": "Fetch User",
                "operationId": "feedf2402b02770f70808b6077657cb2",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "id": "9d12795b-138b-47c9-912e-01eff34cdaa5",
                                        "name": "User",
                                        "email": "user@example.com",
                                        "active": true,
                                        "created_at": "2024-09-22T17:38:29.000000Z",
                                        "updated_at": "2024-09-22T17:38:29.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update User",
                "description": "Update User",
                "operationId": "68beb0720b70b04c958a26a5daf0bf0d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "current_password": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "id": "9d12795b-138b-47c9-912e-01eff34cdaa5",
                                        "name": "User",
                                        "email": "user@example.com",
                                        "active": true,
                                        "created_at": "2024-09-22T17:38:29.000000Z",
                                        "updated_at": "2024-09-22T17:38:29.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create User",
                "description": "Create User",
                "operationId": "c839a51cc6c4dfc6e6c405313abd115d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "id": "9d12795b-138b-47c9-912e-01eff34cdaa5",
                                        "name": "User",
                                        "email": "user@example.com",
                                        "active": true,
                                        "created_at": "2024-09-22T17:38:29.000000Z",
                                        "updated_at": "2024-09-22T17:38:29.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete User",
                "description": "Delete User",
                "operationId": "0439c2b373bd227939e1b4373f0b3760",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
...
}